Started preparation for switching generator to installer
