+++ lib/mix/openmaize.ex
@@ -18,4 +18,22 @@ defmodule Mix.Openmaize do
       Mix.Generator.create_file target, contents
     end
   end
+
+  @doc """
+  Get the mix project name - as a string.
+  """
+  def base_name do
+    Mix.Project.config |> Keyword.fetch!(:app) |> to_string
+  end
+
+  @doc """
+  Get the timestamp for the migration file name.
+  """
+  def timestamp do
+    {{y, m, d}, {hh, mm, ss}} = :calendar.universal_time()
+    "#{y}#{pad(m)}#{pad(d)}#{pad(hh)}#{pad(mm)}#{pad(ss)}"
+  end
+
+  defp pad(i) when i < 10, do: << ?0, ?0 + i >>
+  defp pad(i), do: to_string(i)
 end
