+++ priv/templates/phoenix/phx_api/auth.ex
