+++ priv/templates/phoenix/phx_api/password_reset_view.ex
@@ -1,5 +1,5 @@
 defmodule <%= base %>.PasswordResetView do
-  use <%= base %>.Web, :view<%= if api do %>
+  use <%= base %>.Web, :view
 
   def render("error.json", %{error: message}) do
     %{errors: %{detail: message}}
@@ -7,5 +7,5 @@ defmodule <%= base %>.PasswordResetView do
 
   def render("info.json", %{info: message}) do
     %{info: %{detail: message}}
-  end<% end %>
+  end
 end
