+++ priv/templates/phoenix/phx_api/session_view.ex
@@ -1,7 +1,7 @@
 defmodule <%= base %>.SessionView do
-  use <%= base %>.Web, :view<%= if api do %>
+  use <%= base %>.Web, :view
 
   def render("info.json", %{info: message}) do
     %{info: %{detail: message}}
-  end<% end %>
+  end
 end
