+++ priv/templates/phoenix/phx_html/password_reset_view.ex
@@ -1,0 +1,3 @@
+defmodule <%= base %>.PasswordResetView do
+  use <%= base %>.Web, :view
+end
