+++ priv/templates/phoenix/phx_html/router.ex
@@ -1,22 +1,6 @@
 defmodule <%= base %>.Router do
   use <%= base %>.Web, :router
-<%= if api do %>
-  import <%= base %>.Auth
 
-  pipeline :api do
-    plug :accepts, ["json"]
-    plug :verify_token
-  end
-
-  scope "/api", <%= base %> do
-    pipe_through :api
-
-    post "/sessions/create", SessionController, :create
-    resources "/users", UserController, except: [:new, :edit]<%= if confirm do %>
-    get "/sessions/confirm_email", SessionController, :confirm_email
-    resources "/password_resets", PasswordResetController, only: [:new, :create, :edit, :update]<% end %>
-  end
-<% else %>
   pipeline :browser do
     plug :accepts, ["html"]
     plug :fetch_session
@@ -36,5 +20,4 @@ defmodule <%= base %>.Router do
     get "/sessions/confirm_email", SessionController, :confirm_email
     resources "/password_resets", PasswordResetController, only: [:new, :create, :edit, :update]<% end %>
   end
-<% end %>
 end
