+++ priv/templates/phoenix/phx_html/session_controller_test.exs
@@ -8,29 +8,10 @@ defmodule <%= base %>.SessionControllerTest do
 
   @valid_attrs %{username: "robin", password: "mangoes&g0oseberries"}
   @invalid_attrs %{username: "robin", password: "maaaangoes&g00zeberries"}
-<%= if api do %>
-  setup %{conn: conn} do
-    user = add_user("robin")
-    conn = conn |> add_token_conn(user)
-    {:ok, %{conn: conn, user: user}}
-  end
-
-  test "login succeeds", %{conn: conn} do
-    conn = post conn, session_path(conn, :create), session: @valid_attrs
-    assert json_response(conn, 200)["info"]["detail"]
-  end
-
-  test "login fails", %{conn: conn} do
-    conn = post conn, session_path(conn, :create), session: @invalid_attrs
-    assert json_response(conn, 401)["errors"]["detail"] =~ "need to login"
-  end<%= if confirm do %>
-
-  test "confirmation succeeds for correct key", %{conn: conn} do
-  end
 
   test "confirmation fails for incorrect key", %{conn: conn} do
   end<% end %>
-<% else %>
+
   setup %{conn: conn} do
     conn = conn |> bypass_through(<%= base %>.Router, :browser) |> get("/")<%= if confirm do %>
 
@@ -78,5 +59,4 @@ defmodule <%= base %>.SessionControllerTest do
     assert conn.private.phoenix_flash["error"] =~ "failed"
     assert redirected_to(conn) == session_path(conn, :new)
   end<% end %>
-<% end %>
 end
