+++ priv/templates/phoenix/phx_html/session_view.ex
@@ -1,0 +1,3 @@
+defmodule <%= base %>.SessionView do
+  use <%= base %>.Web, :view
+end
