+++ priv/templates/phoenix/phx_html/user_index.html.eex
