+++ priv/templates/phoenix/phx_html/user_view.ex
@@ -1,0 +1,3 @@
+defmodule <%= base %>.UserView do
+  use <%= base %>.Web, :view
+end
