+++ priv/templates/phoenix/user_model_test.exs
